requirements.txt
pyautogen>=0.2.0
openai>=1.12.0
pandas>=2.0.0
sqlalchemy>=2.0.0
pytest>=7.0.0
requests>=2.31.0 